Resolved page.tsx errors in account-settings for breadcrumbs.
